+++ CHANGELOG.md
@@ -20,7 +20,7 @@
 - Fixed hex word parsing to guard against missing 0x prefix ([#2245](https://github.com/0xMiden/miden-vm/pull/2245)).
 - Systematized u32-indexed vectors ([#2254](https://github.com/0xMiden/miden-vm/pull/2254)).
 - Introduce a new `build_trace()` which builds the trace in parallel from trace fragment contexts ([#1839](https://github.com/0xMiden/miden-vm/pull/1839)) ([#2188](https://github.com/0xMiden/miden-vm/pull/2188))
-- Place the `FastProcessor` stack on the heap instead of the (OS thread) stack (#[2270](https://github.com/0xMiden/miden-vm/pull/2270))
+- Place the `FastProcessor` stack on the heap instead of the (OS thread) stack (#[2271](https://github.com/0xMiden/miden-vm/pull/2271))
  
 ## 0.18.1 (2025-10-02)